+++ tests/UnitTests/PhasorDynamics/runBranchTests.cpp
@@ -9,9 +9,13 @@ int main()
   GridKit::Testing::BranchTests<double, size_t> test;
 
   result += test.constructor();
-  result += test.accessors();
+  result += test.parameterValidation();
+  result += test.dataConstructorDefaults();
+  result += test.parameterSetters();
   result += test.residual();
+  result += test.offNominalResidual();
   result += test.jacobian();
+  result += test.offNominalJacobian();
 
   return result.summary();
 }